Desktop/myinstaapp/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg              
psycopg2-binary
redis
celery
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose